solve migration integers to bigintegers

+++ database/migrations/2025_02_11_002433_create_doctors_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string("nombre");
             $table->string("apellido");
-            $table->integer("telefono");
+            $table->bigInteger("telefono");
             $table->string("especialidad");
             $table->timestamps();
         });
